Intercept all key generation requests with challenges
Change the primary condition for faking an attestation from checking for the ATTEST_KEY purpose to checking for the presence of an attestationChallenge.

This is necessary because apps could request an attestation for general-purpose SIGN keys by providing a challenge during generation. The previous logic incorrectly let these requests pass through to the real TEE, causing the bypass to fail.